requirements.txt
flask==3.1.2
flask-socketio==5.5.1
google-api-python-client==2.179.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
plotly==6.3.0
dash==3.2.0
schedule==1.2.2
python-dotenv==1.1.1
pandas==2.3.2
openpyxl==3.1.5